logger/requirements.txt
paho-mqtt==1.6.1
pyyaml==5.3.1
influxdb-client==1.37.0
